+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# 2.3.0
+
+<!-- Release notes generated using configuration in .github/release.yml at release/2.3.0 -->
+
+## What's Changed
+### Bugfixes
+* bug-fix: include currency-greater-than param for 0 value by @shiqizng in https://github.com/algorand/java-algorand-sdk/pull/620
+### Enhancements
+* api: Sync client object with latest spec. by @winder in https://github.com/algorand/java-algorand-sdk/pull/648
+
+
+**Full Changelog**: https://github.com/algorand/java-algorand-sdk/compare/2.2.0...2.3.0
+
 # 2.2.0
 
 ## Enhancements